moved additional constants and enums from service module to service-api
