+++ service-api/src/main/java/greencity/constant/constant/QuartzConstants.java
@@ -1,4 +1,4 @@
-package greencity.constant;
+package greencity.constant.constant;
 
 import lombok.experimental.UtilityClass;
