+++ service-api/src/main/java/greencity/constant/constant/TariffLocation.java
@@ -1,4 +1,4 @@
-package greencity.constant;
+package greencity.constant.constant;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
