+++ service-api/src/main/java/greencity/constant/constant/pdf/PdfOrderContentDetailsHeaders.java
@@ -1,5 +1,5 @@
 
-package greencity.constant.pdf;
+package greencity.constant.constant.pdf;
 
 import greencity.constant.AppConstant;
 import lombok.Getter;
