+++ service-api/src/main/java/greencity/constant/constant/pdf/PdfQrCodeText.java
@@ -1,4 +1,4 @@
-package greencity.constant.pdf;
+package greencity.constant.constant.pdf;
 
 import greencity.constant.AppConstant;
 import java.util.Locale;
