+++ service-api/src/main/java/greencity/constant/constant/pdf/PdfUnitsOfMeasurement.java
@@ -1,4 +1,4 @@
-package greencity.constant.pdf;
+package greencity.constant.constant.pdf;
 
 import greencity.constant.AppConstant;
 import lombok.Getter;
