+++ service/src/main/java/greencity/service/ubs/AddressServiceImpl.java
@@ -20,9 +20,9 @@ import static java.util.stream.Collectors.toMap;
 import com.google.maps.model.LatLng;
 import greencity.constant.AppConstant;
 import greencity.constant.ErrorMessage;
-import greencity.constant.KyivTariffLocation;
+import greencity.constant.constant.KyivTariffLocation;
 import greencity.constant.OrderHistory;
-import greencity.constant.TariffLocation;
+import greencity.constant.constant.TariffLocation;
 import greencity.dto.CreateAddressRequestDto;
 import greencity.dto.address.AddressDto;
 import greencity.dto.address.UpdateAddressDto;
