+++ service/src/main/java/greencity/ubstelegrambot/service/UserUpdateProcessor.java
@@ -1,6 +1,6 @@
 package greencity.ubstelegrambot.service;
 
-import greencity.constant.TelegramBotConstants;
+import greencity.constant.constant.TelegramBotConstants;
 import greencity.entity.telegram.*;
 import greencity.enums.ChatState;
 import greencity.enums.MessageType;
