+++ service/src/test/java/greencity/ubstelegrambot/factory/keyboards/KeyboardFactoryTest.java
@@ -1,6 +1,6 @@
 package greencity.ubstelegrambot.factory.keyboards;
 
-import greencity.constant.TelegramBotConstants;
+import greencity.constant.constant.TelegramBotConstants;
 import greencity.ubstelegrambot.keyboards.KeyboardFactory;
 import greencity.ubstelegrambot.messages.MessageProvider;
 import org.junit.jupiter.api.Test;
